pyproject.toml
[tool.poetry]
name = "backoffice"
version = "0.1.0"
description = ""
authors = ["Willem Bressers <dhr.bressers@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.2.2"
requests = "^2.33.1"
pandas = "^3.0.1"
reportlab = "4.2.0"
xlsxwriter = "^3.2.9"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
